pyproject.toml
[project]
name = "google-adk-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "black>=25.1.0",
    "google-adk>=0.4.0",
    "litellm>=1.67.6",
]
